Refactor wallet services
