+++ src/main/kotlin/com/ampnet/walletservice/service/WalletService.kt
@@ -17,5 +17,4 @@ interface WalletService {
     fun createOrganizationWallet(organization: UUID, signedTransaction: String, coop: String): Wallet
     fun generatePairWalletCode(publicKey: String): PairWalletCode
     fun getPairWalletCode(code: String): PairWalletCode?
-    fun activateAdminWallet(address: String, coop: String, hash: String): Wallet
 }
