+++ src/test/kotlin/com/ampnet/walletservice/grpc/GrpcWalletServerTest.kt
@@ -9,7 +9,7 @@ import com.ampnet.walletservice.proto.GetWalletsByHashRequest
 import com.ampnet.walletservice.proto.GetWalletsByOwnerRequest
 import com.ampnet.walletservice.proto.WalletResponse
 import com.ampnet.walletservice.proto.WalletsResponse
-import com.ampnet.walletservice.service.WalletService
+import com.ampnet.walletservice.service.CooperativeWalletService
 import io.grpc.stub.StreamObserver
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
@@ -21,7 +21,7 @@ class GrpcWalletServerTest : TestBase() {
     private val coop = "ampnet"
 
     private val mockedWalletRepository = Mockito.mock(WalletRepository::class.java)
-    private val mockedWalletService = Mockito.mock(WalletService::class.java)
+    private val mockedCooperativeWalletService = Mockito.mock(CooperativeWalletService::class.java)
 
     private lateinit var grpcServer: GrpcWalletServer
     private lateinit var testContext: TestContext
@@ -29,7 +29,7 @@ class GrpcWalletServerTest : TestBase() {
     @BeforeEach
     fun init() {
         Mockito.reset(mockedWalletRepository)
-        grpcServer = GrpcWalletServer(mockedWalletRepository, mockedWalletService)
+        grpcServer = GrpcWalletServer(mockedWalletRepository, mockedCooperativeWalletService)
         testContext = TestContext()
     }
